All elements of the AnimationTab have been coded in separate classes. Panels that are not standard communicate through events, to enable a flexible interface that is dependent on the panels present. An example is handling of shift/mouse click or ctrl/mouse click for specific applications or when certain panels are present.
InputParametersTab displays the input parameters in a ScrollPane.

Renderable2DComparator compares on long values, while the implementation of a Renderable2D can generate the value, or use an int internally to conserve memory.

The organization of the dsol-swing components has been severely tightened with a very clear hierarchy: The core DSOLApplication class contains a DSOLPanel for the interaction. DSOLPanel consists of two components: 
(1) at the top a ControlPanel that contains the ClockSpeedPanel, ControlButtonsPanel, RunUntilPanel, and possible other controls. 
(2) A TabbedContentPane that can have e.g., console tabs, statistics tabs, an InputParametersTab, and a DSOLAnimationTab.
The DSOLAnimationTab consists of a AnimationSettingsPanel at the top, a TogglePanel on the left, a PropertiesPanel on the right, and an AnimationPanel at the center. Of all these panels, different versions can be created and linked to the component one level up. One example is the AutoPanAnimationPanel that can track a simulation object. This now just replaces the center component in the DSOLAnimationTab. All panels are accessible for making adaptations.

+++ trunk/dsol-animation-gis/src/main/java/nl/tudelft/simulation/dsol/animation/gis/D2/GisRenderable2D.java
@@ -257,7 +257,7 @@ public class GisRenderable2D implements Renderable2DInterface<GisRenderable2D>, 
 
     /** {@inheritDoc} */
     @Override
-    public int getId()
+    public long getId()
     {
         return -1; // drawn before the rest in case all z-values are the same
     }
