+++ trunk/dsol-animation/src/main/java/nl/tudelft/simulation/dsol/animation/D2/Renderable2DComparator.java
@@ -3,6 +3,7 @@ package nl.tudelft.simulation.dsol.animation.D2;
 import java.util.Comparator;
 
 /**
+ * A comparator for two Renderable2D objects. 
  * <p>
  * Copyright (c) 2002-2021 Delft University of Technology, Jaffalaan 5, 2628 BX Delft, the Netherlands. All rights reserved. See
  * for project information <a href="https://simulation.tudelft.nl/" target="_blank"> https://simulation.tudelft.nl</a>. The DSOL
@@ -45,6 +46,6 @@ public class Renderable2DComparator implements Comparator<Renderable2DInterface<
             // ignore as this can happen when the source is in the process of deletion
             // and therefore it cannot return a proper location.
         }
-        return Integer.compare(r1.getId(), r2.getId());
+        return Long.compare(r1.getId(), r2.getId());
     }
 }
